+++ src/app/app-material.module.ts
@@ -1,24 +1,24 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import {MatGridListModule} from '@angular/material/grid-list';
-import {MatCardModule} from '@angular/material/card';
-import {MatSidenavModule} from '@angular/material/sidenav';
-import {MatToolbarModule} from '@angular/material/toolbar';
-import {MatListModule} from '@angular/material/list';
-import {MatIconModule} from '@angular/material/icon';
-import {MatButtonModule} from '@angular/material/button';
-import {MatDialogModule} from '@angular/material/dialog';
-import {MatSnackBarModule} from '@angular/material/snack-bar';
-//import {MatButtonToggleModule} from '@angular/material/button-toggle';
-import {MatRippleModule} from '@angular/material/core';
-import {MatTableModule} from '@angular/material/table';
-// import {MatFormFieldModule} from '@angular/material/form-field';
-// import {MatInputModule} from '@angular/material/input';
-// import {MatSelectModule} from '@angular/material/select';
-// import {MatSortModule} from '@angular/material/sort';
-import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
-// import {MatCheckboxModule} from '@angular/material/checkbox';
-// import {MatSliderModule} from '@angular/material/slider';
+import { MatGridListModule } from '@angular/material/grid-list';
+import { MatCardModule } from '@angular/material/card';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatListModule } from '@angular/material/list';
+import { MatIconModule } from '@angular/material/icon';
+import { MatButtonModule } from '@angular/material/button';
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
+//import { MatButtonToggleModule } from '@angular/material/button-toggle';
+import { MatRippleModule } from '@angular/material/core';
+import { MatTableModule } from '@angular/material/table';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { MatSelectModule } from '@angular/material/select';
+// import { MatSortModule } from '@angular/material/sort';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+// import { MatCheckboxModule } from '@angular/material/checkbox';
+// import { MatSliderModule } from '@angular/material/slider';
 
 @NgModule({
   declarations: [],
@@ -36,9 +36,9 @@ import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
     MatIconModule,
     MatRippleModule,
     MatTableModule,
-    // MatFormFieldModule,
-    // MatInputModule,
-    // MatSelectModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatSelectModule,
     // MatSortModule,
     MatProgressSpinnerModule,
     // MatSliderModule,
@@ -55,9 +55,9 @@ import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
     MatIconModule,
     MatRippleModule,
     MatTableModule,
-    // MatFormFieldModule,
-    // MatInputModule,
-    // MatSelectModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatSelectModule,
     // MatSortModule,
     MatProgressSpinnerModule,
     // MatSliderModule,