+++ src/app/app-routing.module.ts
@@ -3,6 +3,8 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { IFComponent }   from './if/if.component';
 //import { IFDetailComponent }   from './if-detail/if-detail.component';
+import { VoteComponent }      from './vote/vote.component';
+import { NFTComponent }      from './nft/nft.component';
 import { AboutComponent }      from './about/about.component';
 import { PageNotFoundComponent }      from './page-not-found/page-not-found.component';
 
@@ -16,6 +18,16 @@ const routes: Routes = [
 		path: 'if',
 		component: IFComponent,
 		data: {title: "Home | Internet Freedom"}
+	},
+	{
+		path: 'vote',
+		component: VoteComponent,
+		data: {title: "Vote | Internet Freedom"}
+	},
+	{
+		path: 'nft',
+		component: NFTComponent,
+		data: {title: "NFT | Internet Freedom"}
 	},
 	{
 		path: 'about',