+++ src/app/models/vote.ts
@@ -1,0 +1,10 @@
+export class Vote {
+    account: string;
+    balance: string;
+    address: string;
+    owner: string;
+    delegate: string;
+    choice: string;
+    proposalID: string;
+    reason: string;
+}