+++ src/polyfills.ts
@@ -1,7 +1,7 @@
 (window as any).global = window;
 
-//global.Buffer = global.Buffer || require('buffer').Buffer;
-//global.process = require('process');
+global.Buffer = global.Buffer || require('buffer').Buffer;
+global.process = require('process');
 
 /**
  * This file includes polyfills needed by Angular and is loaded before the app.