+++ .github/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 20221016
+
+* Added vote and nft components
+* Built out web3 integration with Block Native
+* created new models/data for votes and proposals
+* added env variables on Fleek
+* addeded/updated packages for web3 integration
+
 ## 20220922
 
 * First commit